refactors hierarchy for reuse of Simple response
After this first round of implementing commands,
the idea of having a response class for each
command seemed a bit on the heavy side.  All but
two responses have NO data...only an OK or ACK
response.  These all can be handled by a single
Simple.Response class.

This eliminates response classes that use the
names of their command, like Ping.Response and
Repeat.Response.  Maybe the use pattern is tiny
bit more complicated, but the code base is smaller,
and with less duplication.