refactors the SimpleResponse to isolate success/failure
The SimpleResponse base class struck me as a bit smelly.
It combined two ideas:
1) the success/fail of any command,
2) a response without any content other than #1.

There was a mental struggle for non-simple responses
to use 'simple response' as a base class.

non-simple responses needed a base class with
only the success/fail methods.  So does simple response.
Introduced a base class: HealthResponse for all Responses,
simple or not.

rendered the SimpleResponse to be final.  Really,
don't extend SimpleResponse because by logical
definition, the extenstion is not simple!.
(i.e., more complex than SimpleResponse!)
Extend HealthResponse because that is the need.